support for more dead URLs, canonical and absolute path
Form in advsearch now submits to absolute path. Seems Google was
generating paths based on JS change of form’s ‘action’ value for detail
pages. And in some cases this was resulting in paths that did not exist
because of the relative path in the form JS.
support for missing files with pattern agate_anklets.cfm (they can’t be
301ed).
Fixed title  to not to end in ‘in     ‘.  Suffixed gem stone to it. AT
times becomes a repeat of gem stones.

Added private method to backing_bean of gallery to figure out canonical
for several cases